--- kdt-spring-order/src/main/java/org/prgrms/kdt/domain/order/OrderRepository.java
@@ -1,5 +1,0 @@
-package org.prgrms.kdt.domain.order;
-
-public interface OrderRepository {
-    void insert(Order order);
-}
