+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/order/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt.domain.order;
+package org.prgrms.kdt.domain.order.domain;
 
 public enum OrderStatus {
     ACCEPTED,
