+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/order/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package org.prgrms.kdt.domain.order.repository;
+
+import org.prgrms.kdt.domain.order.domain.Order;
+
+public interface OrderRepository {
+    void insert(Order order);
+}
