+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/voucher/repository/VoucherRepository.java
@@ -1,4 +1,6 @@
-package org.prgrms.kdt.domain.voucher;
+package org.prgrms.kdt.domain.voucher.repository;
+
+import org.prgrms.kdt.domain.voucher.domain.Voucher;
 
 import java.util.Optional;
 import java.util.UUID;
