+++ kdt-spring-order/src/main/java/org/prgrms/kdt/domain/voucher/service/VoucherService.java
@@ -1,4 +1,7 @@
-package org.prgrms.kdt.domain.voucher;
+package org.prgrms.kdt.domain.voucher.service;
+
+import org.prgrms.kdt.domain.voucher.domain.Voucher;
+import org.prgrms.kdt.domain.voucher.repository.VoucherRepository;
 
 import java.util.UUID;
