+++ kdt-spring-order/src/test/java/org/prgrms/kdt/OrderTester.java
@@ -1,7 +1,7 @@
 package org.prgrms.kdt;
 
-import org.prgrms.kdt.domain.order.OrderItem;
-import org.prgrms.kdt.domain.order.OrderService;
+import org.prgrms.kdt.domain.order.domain.OrderItem;
+import org.prgrms.kdt.domain.order.service.OrderService;
 import org.prgrms.kdt.infrastructure.configuration.AppConfiguration;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.util.Assert;
